// idcs_portal/src/utils/customInput/CustomInput.types.ts
export interface CustomInputOption {
  name: string
  dropSelect?: React.ReactElement | string | null
  defaultChecked?: boolean
  value?: number | string | null
  subTitleHtml?: React.ReactElement | string | null
  feedbackHtml?: React.ReactElement | string | null
  onChanged?: (event: CustomInputOnChangeEvent) => void
  disabled?: boolean
  hidden?: boolean
}
// ...
export interface CustomInputOnChangeEvent {
  target: {
    value: string | number | null | undefined | boolean
  }
}